final review of the first functional version
