+++ output/1_3PAS_1/1_3PAS_1_EMout.Txt
@@ -72,4 +72,4 @@ Reg 67 = Aux25
 
 Meter           Reg 1      Reg 2      Reg 3      Reg 4      Reg 5      Reg 6      Reg 7      Reg 8      Reg 9      Reg 10     Reg 11     Reg 12     Reg 13     Reg 14     Reg 15     Reg 16     Reg 17     Reg 18     Reg 19     Reg 20     Reg 21     Reg 22     Reg 23     Reg 24     Reg 25     Reg 26     Reg 27     Reg 28     Reg 29     Reg 30     Reg 31     Reg 32     Reg 33     Reg 34     Reg 35     Reg 36     Reg 37     Reg 38     Reg 39     Reg 40     Reg 41     Reg 42     Reg 43     Reg 44     Reg 45     Reg 46     Reg 47     Reg 48     Reg 49     Reg 50     Reg 51     Reg 52     Reg 53     Reg 54     Reg 55     Reg 56     Reg 57     Reg 58     Reg 59     Reg 60     Reg 61     Reg 62     Reg 63     Reg 64     Reg 65     Reg 66     Reg 67  
 
-busa4-_202212598_segmmt_36960       229         64        229        238        214         91        214        233          0          0         49         35         16        -27         16         27          3        -27         13         -0          3         13          2         14          1          0          1          1          0          0          0          0         15          1          0          0          0          0          0          1          1          0          0          0          0          0          1          0          0          0          0          0          0         13          0          0          0          0          0          0          4        210          0          0          0          0          0 
+busa4-_202212598_segmmt_36960       225         60        225        233        209         87        209        226          0          0         35         29         16        -27         16         27          2        -27         13         -0          2         13          2         14          1          0          1          1          0          0          0          0         15          1          0          0          0          0          0          1          1          0          0          0          0          0          1          0          0          0          0          0          0         13          0          0          0          0          0          0          4        205          0          0          0          0          0
