+++ output/1_3PAS_1/Master_DO_01_1_3PAS_1.dss
@@ -7,20 +7,12 @@ Redirect "Trecho_BT_1_3PAS_1.dss"
 Redirect "Ramais_BT_1_3PAS_1.dss"
 Redirect "Transformador_MTBT_1_3PAS_1.dss"
 Redirect "CurvaCarga_1_3PAS_1.dss"
-Redirect 'Medidores.dss'
 Redirect "Cargas_BT_DO01_1_3PAS_1.dss"
 Redirect "Cargas_IP_DO01_1_3PAS_1.dss"
 Redirect "Cargas_MT_DO01_1_3PAS_1.dss"
-Redirect "Tensoesbase_2.dss"
-!Set mode = daily
+Set mode = daily
 Set tolerance = 0.0001
 Set maxcontroliter = 10
 !Set algorithm = newton
 !Solve mode = direct
-batchedit load..* vminpu=0.7
-batchedit load..* vmaxpu=1.2
-batchedit load..* model=1
-Set controlmode=off
 Solve
-Sample
-Show Meters
